% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOS2022BergenStyletable.R
\docType{data}
\name{KOSMOS2022BergenStyletable}
\alias{KOSMOS2022BergenStyletable}
\title{Definition of the colour and style template to be used for plotting}
\format{
An object of class \code{data.frame} with 10 rows and 7 columns.
}
\usage{
KOSMOS2022BergenStyletable
}
\description{
Definition of the colour and style template to be used for plotting
}
\keyword{datasets}
