% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOS2023HelgolandStyletable.R
\docType{data}
\name{KOSMOS2023HelgolandStyletable}
\alias{KOSMOS2023HelgolandStyletable}
\title{Definition of the colour and style template to be used for plotting}
\format{
An object of class \code{data.frame} with 12 rows and 7 columns.
}
\usage{
KOSMOS2023HelgolandStyletable
}
\description{
Definition of the colour and style template to be used for plotting
}
\keyword{datasets}
