% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOS2024KielQuartzSideExperimentStyletable.R
\docType{data}
\name{KOSMOS2024KielQuartzSideExperimentStyletable}
\alias{KOSMOS2024KielQuartzSideExperimentStyletable}
\title{Definition of the colour and style template to be used for plotting}
\format{
An object of class \code{data.frame} with 6 rows and 7 columns.
}
\usage{
KOSMOS2024KielQuartzSideExperimentStyletable
}
\description{
Definition of the colour and style template to be used for plotting
}
\keyword{datasets}
