% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOS2024KielStyletable.R
\docType{data}
\name{KOSMOS2024KielStyletable}
\alias{KOSMOS2024KielStyletable}
\title{Definition of the colour and style template to be used for plotting}
\format{
An object of class \code{data.frame} with 12 rows and 4 columns.
}
\usage{
KOSMOS2024KielStyletable
}
\description{
Definition of the colour and style template to be used for plotting
}
\keyword{datasets}
