% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOSadjustColumnnames.R
\name{KOSMOSadjustColumnames}
\alias{KOSMOSadjustColumnames}
\title{Function to adjust data tables in a way that even slightly mismatching column names can be plotted correctly within this package.}
\usage{
KOSMOSadjustColumnames(
  dataset,
  required_columns = c("Day", "Mesocosm", KOSMOScurrentCategoricalVar, "Delta_TA",
    "Treat_Meso")
)
}
\arguments{
\item{dataset}{A data set object roughly following the common KOSMOS layout for which the column names shall be adjusted to fit the template.}

\item{required_columns}{Vector of those columns to be matched, all others are left untouched. Defaults to all potentially required columns for the available plot functions. This argument is supplied by the plotting function if this function is called automatically, so that only strictly necessary ones are marked as required.}
}
\value{
The same data set with (potentially) modified column names.
}
\description{
Called by either the user or the plotting functions directly, this tool tries to guess the names of the required columns and thereby trys to make the package work even on data tables that deviate from the official template sheet.
}
