% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOScolumntable.R
\docType{data}
\name{KOSMOScolumntable}
\alias{KOSMOScolumntable}
\title{Definition essential columns of an applicable data set and patterns to recognise them by}
\format{
An object of class \code{data.frame} with 7 rows and 5 columns.
}
\usage{
KOSMOScolumntable
}
\description{
Definition essential columns of an applicable data set and patterns to recognise them by
}
\keyword{datasets}
