% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOScurrentCategoricalVar.R
\docType{data}
\name{KOSMOScurrentCategoricalVar}
\alias{KOSMOScurrentCategoricalVar}
\title{Name of the categorical treatment variable}
\format{
An object of class \code{character} of length 1.
}
\usage{
KOSMOScurrentCategoricalVar
}
\description{
Name of the categorical treatment variable
}
\keyword{datasets}
