% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOScurrentContinuousVar.R
\docType{data}
\name{KOSMOScurrentContinuousVar}
\alias{KOSMOScurrentContinuousVar}
\title{Name of the continuous treatment variable}
\format{
An object of class \code{character} of length 1.
}
\usage{
KOSMOScurrentContinuousVar
}
\description{
Name of the continuous treatment variable
}
\keyword{datasets}
