% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOScurrentControlcol.R
\docType{data}
\name{KOSMOScurrentControlcol}
\alias{KOSMOScurrentControlcol}
\title{Definition of the colour used for drawing the control}
\format{
An object of class \code{character} of length 1.
}
\usage{
KOSMOScurrentControlcol
}
\description{
Definition of the colour used for drawing the control
}
\keyword{datasets}
