% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOScurrentControllty.R
\docType{data}
\name{KOSMOScurrentControllty}
\alias{KOSMOScurrentControllty}
\title{Definition of the linetype used for drawing the control}
\format{
An object of class \code{character} of length 1.
}
\usage{
KOSMOScurrentControllty
}
\description{
Definition of the linetype used for drawing the control
}
\keyword{datasets}
