% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOScurrentStatscols.R
\docType{data}
\name{KOSMOScurrentStatscols}
\alias{KOSMOScurrentStatscols}
\title{Definition of the colour used for drawing the stats lines}
\format{
An object of class \code{character} of length 2.
}
\usage{
KOSMOScurrentStatscols
}
\description{
Definition of the colour used for drawing the stats lines
}
\keyword{datasets}
