% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOScurrentStatslty.R
\docType{data}
\name{KOSMOScurrentStatslty}
\alias{KOSMOScurrentStatslty}
\title{Definition of the line type used for drawing the stats lines}
\format{
An object of class \code{character} of length 2.
}
\usage{
KOSMOScurrentStatslty
}
\description{
Definition of the line type used for drawing the stats lines
}
\keyword{datasets}
