% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOScurrentStyletable.R
\docType{data}
\name{KOSMOScurrentStyletable}
\alias{KOSMOScurrentStyletable}
\title{Definition of the colour and style template to be used for plotting}
\format{
An object of class \code{data.frame} with 12 rows and 4 columns.
}
\usage{
KOSMOScurrentStyletable
}
\description{
Definition of the colour and style template to be used for plotting
}
\keyword{datasets}
