% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOScurrentTreatmentSchedule.R
\docType{data}
\name{KOSMOScurrentTreatmentSchedule}
\alias{KOSMOScurrentTreatmentSchedule}
\title{List of the treatment events during the experiment and how to display them}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 2 columns.
}
\usage{
KOSMOScurrentTreatmentSchedule
}
\description{
List of the treatment events during the experiment and how to display them
}
\keyword{datasets}
