% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOSdailyRegressionWrapper.R
\name{KOSMOSdailyRegressionWrapper}
\alias{KOSMOSdailyRegressionWrapper}
\title{Plot a regression for each sampling day quickly}
\usage{
KOSMOSdailyRegressionWrapper(
  filename = "output.svg",
  dataset = KOSMOStestdata,
  parameter = last(names(dataset)),
  days = NULL,
  target_ratio = NULL,
  ...
)
}
\arguments{
\item{filename}{The path and name as which to save the file. This needs to be an \code{.svg} always.}

\item{dataset}{A data set object following the common KOSMOS layout, i.e. loaded from the standard excel data sheet. If left empty, an example dataset \code{KOSMOStestdata} will be plotted to showcase the function. Check \code{View(KOSMOStestdata)} to compare the required data structure.}

\item{parameter}{The column name of the response variable to be plotted given as a string. Defaults to the last column in the data table.}

\item{days}{Select a vector of sampling days to include in the plot pane. By default, all will be included.}

\item{target_ratio}{Select the target aspect ratio for the panel grid. The algorith will try to create a rectangle slightly wider but as close to that as possible to that. By default, it will be \code{16:9} to fit common powerpoint slides.}

\item{...}{forward any other arguments of \code{KOSMOSregplot} to customise the individual plots further.}
}
\description{
Create a file containing a panel of regression plots for each (or a selection of) sampling days automatically. This function is still quite raw and will always create a .svg file.
}
