% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOSdesignfeatures.R
\docType{data}
\name{KOSMOSdesignfeatures}
\alias{KOSMOSdesignfeatures}
\title{Definition of design aspects beyond the main data, such as controls and treatment lines}
\format{
An object of class \code{list} of length 6.
}
\usage{
KOSMOSdesignfeatures
}
\description{
Definition of design aspects beyond the main data, such as controls and treatment lines
}
\keyword{datasets}
