% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOSformatPvalues.R
\name{KOSMOSformatPvalues}
\alias{KOSMOSformatPvalues}
\title{Function to format p-values for display in KOSMOS plots}
\usage{
KOSMOSformatPvalues(p_value)
}
\arguments{
\item{p_value}{a numeric value between 0 and 1.}
}
\value{
Returns a \code{expression()}-compatible string starting with \code{=} or \code{<}.
}
\description{
This function is called in \code{KOSMOSregplot()} to format p-values for display in the little stat stable.
}
\examples{
KOSMOSformatPvalues(0.0132342)
KOSMOSformatPvalues(0.0013234)
KOSMOSformatPvalues(0.0001323)
KOSMOSformatPvalues(0.0000132)
KOSMOSformatPvalues(0.0000013)
}
