% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOSregplot.R
\name{KOSMOSregplot}
\alias{KOSMOSregplot}
\title{Function to plot a regression line per treatment group}
\usage{
KOSMOSregplot(
  dataset = KOSMOStestdata,
  parameter = dimnames(dataset)[[2]][ncol(dataset)],
  days = FALSE,
  ignore = FALSE,
  ylabel = parameter,
  xlabel = "default",
  startat0 = TRUE,
  headspace = 0.3,
  includeThisInYlimit = FALSE,
  ylimit = FALSE,
  axis.tick = "xy",
  axis.show = "xy",
  statsblocklocation = "topleft",
  daylabellocation = "topright",
  new.plot = TRUE,
  ...
)
}
\arguments{
\item{dataset}{A data set object following the common KOSMOS layout, i.e. loaded from the standard excel data sheet. If left empty, an example dataset \code{KOSMOStestdata} will be plotted to showcase the function. Check \code{View(KOSMOStestdata)} to compare the required data structure.}

\item{parameter}{The column name of the response variable to be plotted given as a string. Defaults to the last column in the data table.}

\item{days}{Data from which day or days should be plotted and included in the regression analysis? If more than one day is selected, a mean value of y across those days is calculated per mesocosm. Supply an integer (\code{7}) or vector (\code{c(5,7,9)}). If set to \code{FALSE} (the default), the last sampling day is plotted.}

\item{ignore}{List one or multiple mesocosm numbers to exclude those from the plot and the regression analysis, i.e. \code{c(1,3,10)}. Consider the implications of an unbalanced design in linear regression!}

\item{ylabel}{The y-axis label to be printed. Defaults to the same value as \code{parameter}.}

\item{xlabel}{The x-axis label to be printed. Currently defaults to \code{"Added alkalinity"}.}

\item{startat0}{Should the y-axis start at 0? Can be \code{TRUE} (the default) or \code{False}.}

\item{headspace}{More space needed above the data lines to accommodate the little stats table (see \code{statsblocklocation}), or to include additional features such as labels? \code{headspace} enlarges the y-axis range by the given factor (i.e. \code{0.3}, the dafault) by setting the upper axis limit to \code{130\%} of the original value.}

\item{includeThisInYlimit}{Set this to any value you want included in the range of the y-axis. If the value anyway falls within the range nothing will change, otherwise the lower or upper end of the Y-axis will be shifted to accommodate it. Can be useful if you wish display certain thresholds or reference values.}

\item{ylimit}{Set a fixed range for the y-axis following the pattern \code{c("lower end", "upper end")}, i.e. \code{c(1,3)}. This overwrites \code{startat0}, \code{headspace}, and \code{includeThisInYlimit}. If set to \code{FALSE} (the default), the range will be defined based on the range of data values.}

\item{axis.tick, axis.show}{These options control whether axis ticks and/or labels are displayed. Each can be set to \code{NA} ("show for none"), \code{"x"} ("show for only the x-axis"), \code{"y"} ("show for only the y-axis"), or \code{"xy"} (show for both; the default option). If only \code{axis.tick} is set for an axis the tick marks appear without labels, if both \code{axis.tick} and \code{axis.label} labels are printed next to the ticks.}

\item{statsblocklocation, daylabellocation}{These define where in the plot the little stats table and the label for the displayed sampling day(s) are placed so that they don't collide with the data. The values default to \code{topleft} and \code{topright}, and each can be replaced with any of \code{left, right, top, bottom,} a combination of those, or \code{center} (see [graphics::legend()]). Consider using these options in combination with \code{headspace} to create a clean-looking arrangement.}

\item{new.plot}{If set to \code{FALSE}, the plot will be plotted on-top of an existing, open plot rather than creating a new one. One can use this option to plot data sets on-top of each other, or to set up a plot window independently of the limitations of this function before adding in the data lines. Created for experimentation primarily, his option is rather unstable and prone to unexpected results.}

\item{...}{Some arguments of the base-R \code{plot}-function can be passed on, such as setting a title or the aspect ratio, as well as further graphics parameters. Please check [base::plot()] and [graphics::par()].}
}
\value{
Directly outputs the plot, either to the plot window or another graphics device (e.g. a \code{.png} file) if a connection has been opened.
}
\description{
Creates a regression plot over the alkalinity gradient with a line per mineral treatment group. It works on an excel datasheet following the common KOSMOS layout, assuming a continuous independent variable and a categorical variable with two factors. The current version is limited to work with the KOSMOS Kiel spring 2024 campaign.
}
\examples{
KOSMOSregplot()

}
