% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOSsetup.R
\name{KOSMOSselect}
\alias{KOSMOSselect}
\title{Set up the KOSMOSplotR package to suit your needs}
\usage{
KOSMOSselect(experiment = "kiel")
}
\arguments{
\item{experiment}{Select the experiment that the data to be plotted stems from. This then pre-selects the style template. Choose from "\code{KOSMOS2024Kiel}" (the default), "\code{KOSMOS2023Helgoland}", or "\code{KOSMOS2022Bergen}". Unique sharthands such as "\code{kiel}" will be recognised. Note that you can manually adjust the style template further by overwriting \code{KOSMOScurrentStyletable} in your environment.}
}
\value{
Returns \code{KOSMOScurrentStyletable} as the style template for the selected experiment, \code{KOSMOScurrentStatscols} as the second-darkest colour from each treatment group for ablining the linear model, and \code{KOSMOScurrentCategoricalVar} as the categorical variable studies in the experiment.
}
\description{
Select the campaign you are working on to automatically apply the suitable style template.
}
\examples{
KOSMOSselect("kiel")

KOSMOScurrentStyletable
KOSMOScurrentCategoricalVar
KOSMOScurrentStatscols

}
\seealso{
To overwrite any of these values manually simply assign a new value to them: \code{KOSMOScurrenStyletable[6,2]="red"}. There are further general design features of the plots (such as colour and symbol of the control) saved in \code{KOSMOSdesignfeatures}, that you can access and overwrite if needed.
}
