% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOStestdata.R
\docType{data}
\name{KOSMOStestdata}
\alias{KOSMOStestdata}
\title{Artificial dataset to showcase the functions within the KOSMOSplotsR package}
\format{
An object of class \code{"data.frame"} as if loaded from a standard KOSMOS excel data sheet
\describe{
 \item{Parameter}{Random, unit-less values for a response variable}
}
}
\usage{
KOSMOStestdata
}
\description{
Artificial dataset to showcase the functions within the KOSMOSplotsR package
}
\examples{

data(KOSMOStestdata)
head(KOSMOStestdata)

}
\references{
This data set was artificially created by Nico Sievers
}
\keyword{dataset}
