% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOStestdata_KIEL.R
\docType{data}
\name{KOSMOStestdata_KIEL}
\alias{KOSMOStestdata_KIEL}
\title{Artificial dataset to showcase the functions within the KOSMOSplotR package}
\format{
An object of class \code{"data.frame"} as if loaded from a standard KOSMOS excel data sheet
\describe{
 \item{Parameter}{Random, unit-less values for a response variable}
}
}
\usage{
KOSMOStestdata_KIEL
}
\description{
Artificial dataset to showcase the functions within the KOSMOSplotR package
}
\examples{

data(KOSMOStestdata_KIEL)
head(KOSMOStestdata_KIEL)

}
\references{
This data set was artificially created by Nico Sievers
}
\keyword{dataset}
