% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOStimeplot.R
\name{KOSMOStimeplot}
\alias{KOSMOStimeplot}
\title{KOSMOStimeplot}
\usage{
KOSMOStimeplot(
  dataset,
  parameter,
  ylabel = parameter,
  xlabel = "Experiment day",
  control = FALSE,
  baseline = FALSE,
  ignore = FALSE,
  stats.show = FALSE,
  stats.days = FALSE,
  stats.ignore = FALSE,
  stats.digits = FALSE,
  stats.location = "bottom",
  stats.meanlabel = c("below", "above"),
  stats.doublespecial = FALSE,
  xlimit = FALSE,
  ylimit = FALSE,
  includeThisInYlimit = FALSE,
  startat0 = FALSE,
  headspace = 0,
  axis.tick = "xy",
  axis.show = "xy",
  copepod.draw = FALSE,
  copepod.position = "top",
  treatment.abline = FALSE,
  new.plot = TRUE,
  ...
)
}
\arguments{
\item{dataset}{A data set object following the common KOSMOS layout}

\item{parameter}{The column name of the parameter to be plotted}

\item{ylabel}{The y-axis label, defaults to the same value as \code{parameter}}
}
\value{
directly outputs the plot, either to the plot window or another graphics device (e.g. a .png file) if a connection has been made
}
\description{
Creates a timeline plot over the sampled days with a line per mesocosm. It works on an excel datasheet following the common KOSMOS layout, assuming a continuous independent variable and a categorical variable with two factors. The current version is limited to work with the KOSMOS Kiel spring 2024 campaign.
}
