% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OAEPIIPStyletable.R
\docType{data}
\name{OAEPIIPStyletable}
\alias{OAEPIIPStyletable}
\title{Definition of the colour and style template to be used for plotting}
\format{
An object of class \code{data.frame} with 9 rows and 7 columns.
}
\usage{
OAEPIIPStyletable
}
\description{
Definition of the colour and style template to be used for plotting
}
\keyword{datasets}
