% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KOSMOStimeplot.R
\name{KOSMOStimeplot}
\alias{KOSMOStimeplot}
\title{Function to plot a timeline graph across sampling days}
\usage{
KOSMOStimeplot(
  dataset = KOSMOStestdata,
  parameter = tail(names(dataset), 1),
  subset_data = FALSE,
  exclude_meso = FALSE,
  exclude_day = FALSE,
  control = "Fjord",
  treatmentgroups_sidebyside = FALSE,
  showControlsBothTimes = TRUE,
  ylabel = parameter,
  xlabel = "Experiment day",
  startat0 = FALSE,
  headspace = 0,
  includeThisInYlimit = 0,
  excludeThisFromYlimit = FALSE,
  ylimit = FALSE,
  xlimit = FALSE,
  treatment.abline = TRUE,
  cleaning.abline = FALSE,
  axis.ticks = "xy",
  axis.values = "xy",
  stats.show = FALSE,
  stats.days = FALSE,
  stats.exclude_meso = FALSE,
  stats.digits = FALSE,
  stats.location = "bottom",
  stats.meanlabel = c("below", "above"),
  stats.doublespecial = FALSE,
  copepod.draw = FALSE,
  copepod.position = "top",
  new.plot = TRUE,
  baseline = FALSE,
  ...
)
}
\arguments{
\item{dataset}{A data set object following the common KOSMOS layout, i.e. loaded from the standard excel data sheet. If left empty, an example dataset \code{KOSMOStestdata} will be plotted to showcase the function. Check \code{View(KOSMOStestdata)} to compare the required data structure.}

\item{parameter}{The column name of the response variable to be plotted given as a string. Defaults to the last column in the data table.}

\item{subset_data}{Subset the data by including or excluding rows that have given values in a specified column. If set to \code{FALSE} (the default), no sub-setting is performed. To subset the data table by columns \code{'columnA'} and \code{'columnB'}, supply the following syntax: \code{subset_data = list( columnA = c("value1","value2") , columnB = c("value A","value B"))}, where the column name is the name of the element of the list and the element is a vector (or single value) that marks all rows you wish to include. Alternatively, values can be excluded by adding the prefix \code{"not_"} to a column name, such as \code{subset_data = list( not_columnA = c("value1","value2"))}. Here, rows with \code{value1} and \code{value2} are dropped while all other rows remain. Note that \code{exclude_meso} and \code{exclude_day}, as well as \code{xlimit}, are more convenient parameters to exclude sampling days and mesocosms from the plot.}

\item{exclude_meso, exclude_day}{List one or multiple mesocosm or day numbers, respectively, to exclude those from the plot, i.e. \code{c(1,3,10)}.}

\item{control}{A sample that stands out of the experimental design, such as a harbour or fjord sample, and shall be plotted in a separate style. Name the identifier from the \code{"Mesocosm"} or \code{"Treat_Meso"} column. Defaults to \code{"Fjord"}.}

\item{treatmentgroups_sidebyside}{Choose whether the two treatment groups defined by the categorical factor shall be plotted in one graph (\code{"FALSE"}, the default) or side-by-side in separate plots (\code{"TRUE"}). This option is still under development and might cause issues!}

\item{showControlsBothTimes}{If (\code{treatmentgroups_sidebyside="TRUE"}), by default, those control-mesocosms without added alkalinity are plotted in all panels (\code{"TRUE"}), rather than just with their group (\code{"FALSE"}).}

\item{ylabel}{The y-axis label to be printed. Defaults to the same value as \code{parameter}.}

\item{xlabel}{The x-axis label to be printed. Defaults to \code{"Experiment day"}.}

\item{startat0}{Should the y-axis start at 0? Can be \code{TRUE} or \code{False} (the default), which sets it to the lowest value in the data (which may be negative, therefore consider whether \code{includeThisInYlimit=0} is the more suitable option for you).}

\item{headspace}{More space needed above the data lines to include additional features such as labels? \code{headspace} enlarges the y-axis range by the given factor (i.e. \code{0.25}) by setting the upper axis limit to \code{125\%} of the original value. Defaults to \code{0}.}

\item{includeThisInYlimit}{Set this to any value you want included in the range of the y-axis. If the value anyway falls within the range nothing will change, otherwise the lower or upper end of the Y-axis will be shifted to accommodate it. Can be useful if you wish display certain thresholds or reference values, or make sure that zero is always displayed (the default).}

\item{excludeThisFromYlimit}{Just opposite to above, here you can exclude some data from the axis limits calculation. For example, if you see a strong outlier that causes the rest of the data to be hard to see, you can list it here by its \code{Mesocosm} or \code{Treat_Meso} identifier (also if it is the control you want to exclude). Different to \code{exclude_meso} or \code{exclude_day}, the data will still be plotted but can exceed the axis boundary, so that it is clear that some data is not shown, rather than hiding it completely.}

\item{ylimit}{Set a fixed range for the y-axis following the pattern \code{c("lower end", "upper end")}, i.e. \code{c(1,3)}. This overwrites \code{startat0}, \code{headspace}, and \code{includeThisInYlimit}. If set to \code{FALSE} (the default), the range will be defined based on the range of data values.}

\item{xlimit}{Set a fixed range for the x-axis following the pattern \code{c("lower end", "upper end")}.  If set to \code{FALSE} (the default), the range will include all sampling days for which there is data in the table.}

\item{treatment.abline}{Should treatment additions be marked with vertical lines? \code{TRUE} (the default) or \code{False}.}

\item{cleaning.abline}{Should the days of inside-cleaning of the mesocosms be marked with vertical lines? These could for example have impacted sediment parameters. \code{TRUE} or \code{False} (the default).}

\item{axis.ticks, axis.values}{These options control whether axis ticks and/or labels are displayed. Each can be set to \code{NA} ("show for none"), \code{"x"} ("show for only the x-axis"), \code{"y"} ("show for only the y-axis"), or \code{"xy"} (show for both; the default option). If only \code{axis.ticks} is set for an axis the tick marks appear without labels, if both \code{axis.ticks} and \code{axis.values} labels are printed next to the ticks.}

\item{stats.show}{[\code{TEMPORARILY DISABLED}] Choose whether a linear model shall be calculated and the mean values and p-value for the categorical variable displayed (\code{TRUE} or \code{TRUE}, the default). Note that this currently only works if the categorical variable has at least two levels!}

\item{stats.days}{Data from which day or days should be included in the stats analysis? If more than one day is selected, a mean value of y across those days is calculated per mesocosm. Supply an integer (\code{7}) or vector containing the first and last day (\code{c(5,9)}). If set to \code{FALSE} (the default), the last sampling day is plotted.}

\item{stats.exclude_meso}{List one or multiple mesocosms to exclude those from the stats analysis, i.e. \code{c(1,3,10)}. Mesocosms excluded from the plot via \code{exclude_meso} are anyway excluded.}

\item{stats.digits}{The number of digits of the displayed values.}

\item{stats.location}{Position of the stats text given as \code{"top"}, \code{"centre"}, or \code{"bottom"} (the default)}

\item{stats.meanlabel}{Indicate whether the label of a factor level's mean value should be displayed \code{above} or \code{below} the mark, in the format \code{c([lower value],[higher value])}. Defaults to \code{c("below","above")}.}

\item{stats.doublespecial}{\code{(don}'\code{t ask...)}}

\item{copepod.draw}{\code{(outdated, unavailable)}}

\item{copepod.position}{\code{(outdated, unavailable)}}

\item{new.plot}{If set to \code{FALSE}, the plot will be plotted on-top of an existing, open plot rather than creating a new one. One can use this option to plot data sets on-top of each other, or to set up a plot window independently of the limitations of this function before adding in the data lines. Created for experimentation primarily, his option is rather unstable and prone to unexpected results.}

\item{baseline}{\code{(currently unavailable)}}

\item{...}{Some arguments of the base-R \code{plot}-function can be passed on, such as setting a title or the aspect ratio, as well as further graphics parameters. Please check [base::plot()] and [graphics::par()].}
}
\value{
Directly outputs the plot, either to the plot window or another graphics device (e.g. a \code{.png} file) if a connection has been opened.
}
\description{
Creates a timeline plot over the sampled days with a line per mesocosm. It works on an excel datasheet roughly following the common KOSMOS layout, assuming a continuous independent variable and a categorical variable with two factors.
}
